setup footer
